requirements.txt
Flask[async]==3.0.3
pypdf==5.1.0
langchain_chroma==0.1.4
langchain_community==0.3.7
langchain_core==0.3.19
langchain_ollama==0.2.0
langchain_text_splitters==0.3.2
PyYAML==6.0.1
